+++ src/xxhash.rs
@@ -27,7 +27,7 @@ impl XXH32 {
 
     #[inline]
     pub fn hash_with_seed(seed: u32, bytes: &[u8]) -> u32 {
-        unsafe { C::XXH32(bytes.as_ptr() as *const c_void, bytes.len(), seed) }
+        unsafe { C::XXH32(bytes.as_ptr() as *const c_void, bytes.len() as u64, seed) }
     }
 
     #[inline]
@@ -52,7 +52,7 @@ impl XXH32 {
             C::XXH32_update(
                 &mut self.state,
                 bytes.as_ptr() as *const c_void,
-                bytes.len(),
+                bytes.len() as u64,
             );
         }
     }
@@ -83,7 +83,7 @@ impl XXH64 {
 
     #[inline]
     pub fn hash_with_seed(seed: u64, bytes: &[u8]) -> u64 {
-        unsafe { C::XXH64(bytes.as_ptr() as *const c_void, bytes.len(), seed) }
+        unsafe { C::XXH64(bytes.as_ptr() as *const c_void, bytes.len() as u64, seed) }
     }
 
     #[inline]
@@ -110,7 +110,7 @@ impl Hasher for XXH64 {
             C::XXH64_update(
                 &mut self.state,
                 bytes.as_ptr() as *const c_void,
-                bytes.len(),
+                bytes.len() as u64,
             );
         }
     }
